+++ doc/CHANGES.rst
@@ -33,6 +33,7 @@ Future
 * [ADD] Add Color Column
 * [REF] column can be encrypted
 * [REF] DataTime column is not a naive datatime value
+* [ADD] Add Password Column
 
 0.7.2 (2016-01-14)
 ------------------